// ui/lib/types/logs.ts
export type MessageContentType =
	| "text"
	| "image_url"
	| "input_audio"
	| "input_text"
	| "input_file"
	| "output_text"
	| "refusal"
	| "reasoning";

export interface ContentBlock {
	type: MessageContentType;
	text?: string;
	image_url?: {
		url: string;
		detail?: string;
	};
	input_audio?: {
		data: string;
		format?: string;
	};
}

export type ChatMessageContent = string | ContentBlock[];

export interface ChatMessage {
	role: "assistant" | "user" | "system" | "chatbot" | "tool";
	content: ChatMessageContent;
	tool_call_id?: string;
	refusal?: string;
	annotations?: Annotation[];
	tool_calls?: ToolCall[]; // For backward compatibility, tool calls are now in the content
	thought?: string;
}
// ...
export interface FunctionCall {
	name?: string;
	arguments: string; // stringified JSON
}

export interface ToolCall {
	type?: string;
	id?: string;
	function: FunctionCall;
}
// ...
export interface Citation {
	start_index: number;
	end_index: number;
	title: string;
	url?: string;
	sources?: unknown;
	type?: string;
}

export interface Annotation {
	type: string;
	url_citation: Citation;
}